auth_service/tests/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
backoff==2.1.2
pytest==7.1.2
python-dotenv==1.0.0
